requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dotenv==0.9.9
fastapi==0.120.2
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
